fix: CI/CD pipeline failures - frontend & backend fixes
## Changes:
1. **Frontend TypeScript fixes** (Agent UI & Admin UI):
   - Fixed inconsistent query return types in DashboardPage activities query
   - Changed empty Payload interfaces to type aliases to avoid TypeScript errors
   - Extracted non-component functions from context files (React Fast Refresh fix)
   - Added utility files for theme and auth helpers

2. **Prisma 7 compatibility**:
   - Removed `url` from datasource (now in prisma.config.ts)
   - Prisma schema now uses only provider declaration

3. **Build configuration fixes**:
   - Excluded e2e and playwright config from TypeScript build
   - Fixed tsconfig.build.json to properly exclude test files

## Fixes:
- Frontend builds now pass (agent-ui & admin-ui)
- Backend Prisma generation now succeeds
- All type checking passes with strict TypeScript settings
