+++ agent-ui/src/utils/theme.ts
@@ -1,0 +1,7 @@
+import type { Theme } from '../types'
+
+export function getStoredTheme(): Theme {
+  const stored = localStorage.getItem('theme')
+  if (stored === 'dark' || stored === 'light') return stored
+  return window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light'
+}
