+++ prisma/schema.prisma
@@ -3,8 +3,7 @@ generator client {
 }
 
 datasource db {
-  provider  = "postgresql"
-  url       = env("DATABASE_URL")
+  provider = "postgresql"
 }
 
 // ─── Enums ───────────────────────────────────────────────────────────
