+++ tsconfig.build.json
@@ -1,4 +1,4 @@
 {
   "extends": "./tsconfig.json",
-  "exclude": ["node_modules", "test", "dist", "admin-ui", "agent-ui", "mobile", "prisma", "**/*spec.ts"]
+  "exclude": ["node_modules", "test", "dist", "admin-ui", "agent-ui", "mobile", "prisma", "e2e", "playwright.config.ts", "**/*spec.ts"]
 }
